feat(unityengine): fix labels and some d3d11on12 games maybe

+++ src/games/unityengine/Plugin/Scion/CombinationPass_0x2E658124.ps_4_0.hlsl
@@ -124,7 +124,7 @@ void main(
                                                                      : renodx::tonemap::config::hue_correction_type::CUSTOM;
   config.hue_correction_strength = injectedData.toneMapHueCorrection;
   config.hue_correction_color = lerp(r0.xyz, hueCorrectionColor, injectedData.toneMapHueShift);
-  config.reno_drt_hue_correction_method = (int)injectedData.toneMapHueProcessor;
+  config.reno_drt_hue_correction_method = injectedData.toneMapHueProcessor;
   config.reno_drt_tone_map_method = injectedData.toneMapType == 3.f ? renodx::tonemap::renodrt::config::tone_map_method::REINHARD
                                                                     : renodx::tonemap::renodrt::config::tone_map_method::DANIELE;
   config.reno_drt_per_channel = injectedData.toneMapPerChannel != 0.f;
